Fix timezone_select method signature to provide better support for priorities.
time_zone_select() would raise TypeError as written saying 'no implicit
conversion of Hash into Array'. This is because the 2nd parameter to
FormOptionsHelper#time_zone_select is supposed to be an Array. And
FoundationRailsHelper::FormBuilder#time_zone_select() was passing in a Hash.

The array being passed in is to be a list of priority time zones that show up at
the top of the select list. This option is now supported.