(chore): Travis tests multiple rails versions
In our readme we're specifically supporting rails 4.1, 4.2 and 5, so
let's test against those!

This old post by @schneems provides the pattern:
http://schneems.com/post/50991826838/testing-against-multiple-rails-versions/

Testing against master (like in the post) would be useful for flagging
up any upcoming issues, but since we're using the individual gems it's a
bit too fiddly.

Rails 5 only supports ruby 2.2 and up, so we need to exclude the other
ruby versions