(delete): unused html options in extension
These are legacy:

The initialisation of the html options hash was introduced in
3713cd4e923e68234f3d63e3313f3ed1c7c596b0 when a class ('nice') was added
onto all forms.

It could have been removed in #60 when that class was removed. It's no
longer required.

The attached_labels key has never been used and seems to have been added
by mistake in pull request #47