(refactor): auto_labels default in form builder
auto_labels is the flag which determines whether labels are
automatically added to fields based on field name.

Rather than setting a default value twice in the action_view_extension,
we can set it once in the form builder (which is the only place that it
is used).

Using `fetch` gives us exactly the same behaviour as checking for
`has_key?`.

Note that:

- `@options` is the options hash for the _form_
- `options` is the options hash for the _field_