(fix): auto_labels defaults to true
We have decided that `auto_labels` should be true by default. That means:
it should always be true unless explicitly set to false.

In previous implementations if `auto_labels` was explicitly set to nil,
this would be interpreted in the same way as `auto_labels: false`, since
nil is falsy.

This commit fixes that case, and backfills some other test cases around
usage of `auto_labels`.