(refactor): Extract a SizeClassCalculator
...for working out what column size classes to apply, based on options.

This squashes one of the rubocop complaints about high complexity of
the related method, which had 3 separate if statements.

The problem is basically equivalent to handling POST-ed data: the user
has given us some data and we need to sanitize them.

SizeClassCalculator is a method object which handles this. I guess
there's an argument for moving it outside the containing class, but I
don't see any harm in leaving it where it is.

I'm sure there are further refactorings which could be done on this
class, but hopefully this is a good start.