Adding closable keyword option to display_flash_messages
Making a breaking change to display_flash_messages that requires the key_matching hash to be prefixed with the :key_matching keyword.

Adding the keyword option :closable to the display_flash_messages method that accepts a boolean true/false, defaulting to true.

If false, suppresses the close buttons on the callouts.